+++ packages/org/_initialData/locale/uk^020_org_settings.js
@@ -1,0 +1,19 @@
+const loader = require('@unitybase/base').dataLoader
+
+module.exports = function (session) {
+  const localizationConfig = {
+    entity: 'ubs_settings',
+    keyAttribute: 'settingKey',
+    localization: [
+      {
+        keyValue: 'org.organization.OKPORequired',
+        execParams: {
+          name: 'Обов’язковість ЄДРПОУ для організацій',
+          description: 'Встановлює обов’язковість коду ЄДРПОУ при створенні внутрішніх і зовнішніх організацій'
+        }
+      }
+    ]
+  }
+
+  loader.localizeEntity(session, localizationConfig, __filename)
+}